+++ application/models/Tracker/Sites/LHTranslation.php
@@ -5,12 +5,12 @@ class LHTranslation extends Base_Site_Model {
 	public $chapterFormat = '/^[0-9\.]+$/';
 
 	public function getFullTitleURL(string $title_url) : string {
-		return "http://lhtranslation.com/manga-{$title_url}.html";
+		return "http://lhtranslation.com/{$title_url}";
 	}
 
 	public function getChapterData(string $title_url, string $chapter) : array {
 		return [
-			'url'    => "http://lhtranslation.com/read-{$title_url}-chapter-{$chapter}.html",
+			'url'    => "http://read.lhtranslation.com/read-{$title_url}-chapter-{$chapter}.html",
 			'number' => "c{$chapter}"
 		];
 	}
@@ -18,25 +18,21 @@ class LHTranslation extends Base_Site_Model {
 	public function getTitleData(string $title_url, bool $firstGet = FALSE) : ?array {
 		$titleData = [];
 
-		$fullURL = $this->getFullTitleURL($title_url);
-
+		$fullURL = "http://lhtranslation.com/{$title_url}/feed/";
 		$content = $this->get_content($fullURL);
 
-		$data = $this->parseTitleDataDOM(
-			$content,
-			$title_url,
-			"//title",
-			"//div[@id='tab-chapper']/div/ul/table/tbody/tr[1]",
-			"td[2]/i/time",
-			"td[1]/a[1]",
-			"Donate for Us" //bad titles simply get redirected to front page
-		);
-		if($data) {
-			$titleData['title'] = str_replace(' - LHtranslation', '', trim($data['nodes_title']->textContent));
+		$data = $content['body'];
+		$xml = simplexml_load_string($data) or die("Error: Cannot create object");
+		if(((string) $xml->{'channel'}->title) !== 'Comments on: '){
+			if(isset($xml->{'channel'}->item[0])) {
+				$titleData['title'] = trim(substr((string) $xml->{'channel'}->title, 0, -33));
 
-			$titleData['latest_chapter'] = preg_replace('/^read-(?:.*?)chapter-(.*?)\.html$/', '$1', (string) $data['nodes_chapter']->getAttribute('href'));
-
-			$titleData['last_updated'] =  date("Y-m-d H:i:s", strtotime((string) $data['nodes_latest']->textContent));
+				$titleData['latest_chapter'] = str_replace('-', '.', preg_replace('/^.*?-(?:.*?)chapter-(.*?)\/$/', '$1', (string) $xml->{'channel'}->item[0]->link));
+				$titleData['last_updated'] =  date("Y-m-d H:i:s", strtotime((string) $xml->{'channel'}->item[0]->pubDate));
+			}
+		} else {
+			log_message('error', "Series missing? (LHTranslation): {$title_url}");
+			return NULL;
 		}
 
 		return (!empty($titleData) ? $titleData : NULL);